Write data and logs to external storage if connected
If a an external storage device is connected, write the data and
logs to it instead of the SD card. If one isn't connected, it will
write to the data and logs directories in the repo directory.

Also, the log file will be removed at the beginning of every month
to save disk space.

The readme was also updated.